Minor updated localization files
