+++ edge_detector.py
@@ -2,19 +2,13 @@ import matplotlib.pyplot as plt
 import numpy as np
 from gaussian import GaussianImage, nms, hysteresis, thresh
 
-def display_image(img, title="figure"):
-    plt.title(title)
-    plt.imshow(img)  
-    plt.axis('off')      
-    plt.show() 
-
 # # #Canny edge detector steps:
 # # Apply Gaussian filter to smooth the image in order to remove the noise
 # # Find the intensity gradients of the image
 # # Apply gradient magnitude thresholding or lower bound cut-off suppression to get rid of spurious response to edge detection
 # # Apply double threshold to determine potential edges
 # # Track edge by hysteresis: Finalize the detection of edges by suppressing all the other edges that are weak and not connected to strong edges.
-def detect_edges(img):
+def detect_edges_canny(img):
     #1) Apply gaussian filters to x and y
     size = 25
     sigma = 1.4
@@ -41,3 +35,5 @@ def detect_edges(img):
     return final
 
 
+def detect_edges_tight(img):
+    print('Hi')
